--- test/db/_ff_data/rand/_delta_log/00000000000000000000.json
@@ -1,6 +1,0 @@
-{"protocol":{"minReaderVersion":1,"minWriterVersion":2}}
-{"metaData":{"id":"119354ff-c35e-4d96-9dbf-29ac4fbe04b9","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"part\",\"type\":\"string\",\"nullable\":true,\"metadata\":{}},{\"name\":\"value\",\"type\":\"double\",\"nullable\":true,\"metadata\":{}}]}","partitionColumns":["part"],"configuration":{},"createdTime":1654122599448}}
-{"add":{"path":"part=a/part-00000-6afc90b0-393b-46a4-a497-93197a30fd04.c000.snappy.parquet","partitionValues":{"part":"a"},"size":748,"modificationTime":1654122599830,"dataChange":true,"stats":"{\"numRecords\":32,\"minValues\":{\"value\":0.018242481428361046},\"maxValues\":{\"value\":0.9731103492701043},\"nullCount\":{\"value\":0}}"}}
-{"add":{"path":"part=b/part-00000-571a89d4-897f-4183-92ee-4a054ff9e9ef.c000.snappy.parquet","partitionValues":{"part":"b"},"size":780,"modificationTime":1654122599850,"dataChange":true,"stats":"{\"numRecords\":36,\"minValues\":{\"value\":0.0025950243836465603},\"maxValues\":{\"value\":0.9972555347683943},\"nullCount\":{\"value\":0}}"}}
-{"add":{"path":"part=c/part-00000-e705a1df-082c-445a-8822-b4f6fae25765.c000.snappy.parquet","partitionValues":{"part":"c"},"size":748,"modificationTime":1654122599860,"dataChange":true,"stats":"{\"numRecords\":32,\"minValues\":{\"value\":0.00753436313407474},\"maxValues\":{\"value\":0.9750671629451425},\"nullCount\":{\"value\":0}}"}}
-{"commitInfo":{"timestamp":1654122600012,"operation":"WRITE","operationParameters":{"mode":"Overwrite","partitionBy":"[\"part\"]"},"isolationLevel":"Serializable","isBlindAppend":false,"operationMetrics":{"numFiles":"3","numOutputRows":"100","numOutputBytes":"2276"},"engineInfo":"Apache-Spark/3.2.1 Delta-Lake/1.2.1","txnId":"2f244a90-675b-4f91-a08b-40bddb25473e"}}
