--- test/db/_ff_data/simple/_delta_log/00000000000000000000.json
@@ -1,4 +1,0 @@
-{"protocol":{"minReaderVersion":1,"minWriterVersion":2}}
-{"metaData":{"id":"a4ff9293-b23a-4281-a215-71bae9968265","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"col1\",\"type\":\"double\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col2\",\"type\":\"double\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col3\",\"type\":\"double\",\"nullable\":true,\"metadata\":{}}]}","partitionColumns":[],"configuration":{},"createdTime":1654121665343}}
-{"add":{"path":"part-00000-3b23e96c-c274-4c12-805c-766955e71034-c000.snappy.parquet","partitionValues":{},"size":1097,"modificationTime":1654121665650,"dataChange":true,"stats":"{\"numRecords\":5,\"minValues\":{\"col1\":0.24196227156603412,\"col2\":-1.913280244657798,\"col3\":-1.7249178325130328},\"maxValues\":{\"col1\":1.5792128155073915,\"col2\":0.7674347291529088,\"col3\":0.6476885381006925},\"nullCount\":{\"col1\":0,\"col2\":0,\"col3\":0}}"}}
-{"commitInfo":{"timestamp":1654121665673,"operation":"WRITE","operationParameters":{"mode":"ErrorIfExists","partitionBy":"[]"},"isolationLevel":"Serializable","isBlindAppend":true,"operationMetrics":{"numFiles":"1","numOutputRows":"5","numOutputBytes":"1097"},"engineInfo":"Apache-Spark/3.2.1 Delta-Lake/1.2.1","txnId":"4f2a1018-4b2e-4a24-8325-6cc8ad83ec71"}}
