--- test/db/_ff_data/simple/_delta_log/00000000000000000001.json
@@ -1,3 +1,0 @@
-{"add":{"path":"part-00000-4e33d745-6e05-4dd3-8d32-49936e5966ab-c000.snappy.parquet","partitionValues":{},"size":1097,"modificationTime":1654121766690,"dataChange":true,"stats":"{\"numRecords\":5,\"minValues\":{\"col1\":0.24196227156603412,\"col2\":-1.913280244657798,\"col3\":-1.7249178325130328},\"maxValues\":{\"col1\":1.5792128155073915,\"col2\":0.7674347291529088,\"col3\":0.6476885381006925},\"nullCount\":{\"col1\":0,\"col2\":0,\"col3\":0}}"}}
-{"remove":{"path":"part-00000-3b23e96c-c274-4c12-805c-766955e71034-c000.snappy.parquet","deletionTimestamp":1654121767078,"dataChange":true,"extendedFileMetadata":true,"partitionValues":{},"size":1097}}
-{"commitInfo":{"timestamp":1654121767079,"operation":"WRITE","operationParameters":{"mode":"Overwrite","partitionBy":"[]"},"readVersion":0,"isolationLevel":"Serializable","isBlindAppend":false,"operationMetrics":{"numFiles":"1","numOutputRows":"5","numOutputBytes":"1097"},"engineInfo":"Apache-Spark/3.2.1 Delta-Lake/1.2.1","txnId":"fe64d556-b12b-4785-879a-f83b6091ebf7"}}
